finished

+++ ex2/RankTree.cpp
@@ -110,9 +110,9 @@ Employee* RankTree::getElementByReverseRank(int reverse_rank)
     return temp->getData();
 }
 
-int RankTree::sumOfGradeTopWorkers(int m)
+long int RankTree::sumOfGradeTopWorkers(int m)
 {
-    int sum = 0;
+    long int sum = 0;
     RankNode *temp = root;
     if(temp == nullptr || temp->getNumEmployees() < m){
         throw NotEnoughEmployees();
@@ -134,7 +134,7 @@ int RankTree::sumOfGradeTopWorkers(int m)
     throw RankNotFound();
 }
 
-void RankTree::averageGradesInSalaryRange(int lower, int higher, int* num_employees_in_range, int *sum_grades_in_range)
+void RankTree::averageGradesInSalaryRange(int lower, int higher, long int* num_employees_in_range, long int *sum_grades_in_range)
 {
     *num_employees_in_range = root->getNumEmployees();
     *sum_grades_in_range = root->getSumGrades();
