+++ ex2/RankTree.h
@@ -330,8 +330,8 @@ class RankTree
         void remove(Employee* data);
         RankNode* findMinNode(RankNode* start);
         RankNode* findMaxNode(RankNode* start);
-        int sumOfGradeTopWorkers(int m);
-        void averageGradesInSalaryRange(int lower, int higher, int* num_employees_in_range, int *sum_grades_in_range);
+        long int sumOfGradeTopWorkers(int m);
+        void averageGradesInSalaryRange(int lower, int higher, long int* num_employees_in_range, long int* sum_grades_in_range);
         Employee* getElementByReverseRank(int reverse_rank);
         void absorbTree(RankTree& tree);
         void destroyRecursiveData(RankNode* node);
